# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Baslid Project Api Documentation",
        "description": "This is a documentation for the Baslid Project witch is produced by Swagger",
        "contact": {
            "email": "aminmohammadmazreali@gmail.com"
        },
        "license": {
            "name": "Nginx 1.17.7",
            "url": "https://nginx.org/LICENSE"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Base Url"
        }
    ],
    "paths": {
...
        "/auth/upload-avatar": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "upload user avatar",
                "description": "upload user avatar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UploadAvatarUserRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "client_bearer_token": []
                    }
                ]
            }
        },
        "/product-statistic/earning-history": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "get earning history",
                "description": "get earning history for user",
                "parameters": [
                    {
                        "name": "prevDays",
                        "in": "query",
                        "description": "previous days for report",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation when send prevDays",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EarningHistoryForPrevDaysDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "2000": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EarningHistoryForYearDaysDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "client_bearer_token": []
                    }
                ]
            }
        },
...
        "/product-statistic/{userDiscountCode}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "store new product statistic for user",
                "description": "store new product statistic for user",
                "parameters": [
                    {
                        "name": "userDiscountCode",
                        "in": "path",
                        "description": "user discount code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductStatisticRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductStatisticDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "EarningHistoryForPrevDaysDTO": {
                "title": "Get Earning history Dto",
                "description": "Get Earning history data",
                "properties": {
                    "2023-05-1": {
                        "title": "2023-05-1",
                        "type": "integer",
                        "example": 13
                    },
                    "2023-05-2": {
                        "title": "2023-05-2",
                        "type": "integer",
                        "example": 145
                    },
                    "2023-05-3": {
                        "title": "2023-05-3",
                        "type": "integer",
                        "example": 113
                    },
                    "2023-05-4": {
                        "title": "2023-05-4",
                        "type": "integer",
                        "example": 100
                    },
                    "2023-05-5": {
                        "title": "2023-05-5",
                        "type": "integer",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "EarningHistoryForYearDaysDTO": {
                "title": "Get Earning history Dto",
                "description": "Get Earning history data",
                "properties": {
                    "1": {
                        "title": "1",
                        "type": "integer",
                        "example": 223
                    },
                    "2": {
                        "title": "2",
                        "type": "integer",
                        "example": 13
                    },
                    "3": {
                        "title": "3",
                        "type": "integer",
                        "example": 4
                    },
                    "4": {
                        "title": "4",
                        "type": "integer",
                        "example": 54
                    },
                    "5": {
                        "title": "5",
                        "type": "integer",
                        "example": 133
                    },
                    "6": {
                        "title": "6",
                        "type": "integer",
                        "example": 11
                    },
                    "7": {
                        "title": "7",
                        "type": "integer",
                        "example": 43
                    },
                    "8": {
                        "title": "8",
                        "type": "integer",
                        "example": 2
                    },
                    "9": {
                        "title": "9",
                        "type": "integer",
                        "example": 5
                    },
                    "10": {
                        "title": "10",
                        "type": "integer",
                        "example": 6
                    },
                    "11": {
                        "title": "11",
                        "type": "integer",
                        "example": 22
                    },
                    "12": {
                        "title": "12",
                        "type": "integer",
                        "example": 135
                    }
                },
                "type": "object"
            },
...
            "Unauthenticated": {
                "title": "Unauthenticated",
                "description": "Unauthenticated Request",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
...
            "UploadAvatarUserRequestBody": {
                "title": "Upload Avatar User Request Body",
                "description": "Upload Avatar User Request Body",
                "properties": {
                    "avatar": {
                        "title": "avatar",
                        "description": "avatar of user",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
            "StoreProductStatisticRequestBody": {
                "title": "Store Product Statistic Request Body",
                "description": "Store Product Statistic Request Body",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of product",
                        "type": "string",
                        "example": "mouse"
                    },
                    "image": {
                        "title": "image",
                        "description": "image url of product",
                        "type": "string",
                        "example": "mouse.jpeg"
                    },
                    "date": {
                        "title": "date",
                        "description": "date of product",
                        "type": "string",
                        "example": "2023-05-05"
                    },
                    "tendered": {
                        "title": "tendered",
                        "description": "tendered of product",
                        "type": "integer",
                        "example": 200
                    },
                    "earning": {
                        "title": "earning",
                        "description": "earning of product",
                        "type": "integer",
                        "example": 10
                    },
                    "status": {
                        "title": "status",
                        "description": "paid status of product",
                        "type": "boolean",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "ProductStatisticDTO": {
                "title": "Product Statistic Dto",
                "description": "Product Statistic data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of product",
                        "type": "string",
                        "example": "mouse"
                    },
                    "image": {
                        "title": "image",
                        "description": "image url of product",
                        "type": "string",
                        "example": "mouse.jpeg"
                    },
                    "date": {
                        "title": "date",
                        "description": "date of product",
                        "type": "string",
                        "example": "2023-05-05"
                    },
                    "tendered": {
                        "title": "tendered",
                        "description": "tendered of product",
                        "type": "integer",
                        "example": 200
                    },
                    "earning": {
                        "title": "earning",
                        "description": "earning of product",
                        "type": "integer",
                        "example": 10
                    },
                    "status": {
                        "title": "status",
                        "description": "paid status of product",
                        "type": "boolean",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UserDTO": {
                "title": "User Dto",
                "description": "User data",
                "properties": {
                    "firstName": {
                        "title": "firstName",
                        "description": "first name of user",
                        "type": "string",
                        "example": "Amin"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "last name of user",
                        "type": "string",
                        "example": "Mazreali"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of user",
                        "type": "string",
                        "example": "09214125578"
                    },
                    "type": {
                        "title": "type",
                        "description": "type of user",
                        "type": "integer",
                        "example": 2
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "amin@gmail.com"
                    },
                    "country": {
                        "title": "country",
                        "description": "country of user",
                        "type": "string",
                        "example": "iran"
                    },
                    "id": {
                        "title": "id",
                        "description": "id of user",
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "title": "address",
                        "description": "address of user",
                        "type": "string",
                        "example": "Iran, Test"
                    },
                    "avatar": {
                        "title": "avatar",
                        "description": "avatar of user",
                        "type": "string",
                        "example": "test.png"
                    },
                    "language": {
                        "title": "language",
                        "description": "language of user",
                        "type": "string",
                        "example": "Persian"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of user",
                        "type": "string",
                        "example": "Euro"
                    },
                    "biography": {
                        "title": "biography",
                        "description": "biography of user",
                        "type": "string",
                        "example": "test, test"
                    },
                    "paypalAddress": {
                        "title": "paypalAddress",
                        "description": "paypal address of user",
                        "type": "string",
                        "example": "paypal.com"
                    },
                    "token": {
                        "title": "token",
                        "description": "token of user",
                        "type": "string",
                        "example": "jwt token"
                    },
                    "discountCode": {
                        "title": "discountCode",
                        "description": "discount code of user",
                        "type": "string",
                        "example": "1231241"
                    },
                    "discountPercent": {
                        "title": "discountPercent",
                        "description": "discount percent of user",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "client_bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}